tests/wpt/tests/html/browsers/windows/consume-user-activation/window-open.html

<!DOCTYPE html>
<title>window.open() and consuming user activation</title>
<link rel="help" href="https://html.spec.whatwg.org/#the-rules-for-choosing-a-navigable">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';

promise_test(async function(t) {
  await test_driver.bless("user activation");
  const testWin = window.open("/resources/blank.html", "_blank");
  t.add_cleanup(() => {
    testWin.close();
  });
  assert_false(navigator.userActivation.isActive, "User activation should be consumed");
});
</script>
